refactor(visio): extraire useVisioActions + visioTimeWindow de VisioManager (G8)
Extrait la logique temporelle pure (fenêtre -15/+30min, format heure) vers
src/utils/visioTimeWindow.js, et l'état + les 7 appels API (programmer,
désactiver, démarrer, terminer, télécharger PDF, rejoindre, compteur) vers
src/composables/useVisioActions.js. Template, CSS, routes, payloads et
événements émis inchangés ; import side-effect de services/jitsi conservé.
Script du composant 367 -> 99 lignes.
